[REFACTOR] 패키지 및 프로젝트 명 변경 server -> e-commerce
