+++ settings.gradle.kts
@@ -1,1 +1,1 @@
-rootProject.name = "server"
+rootProject.name = "e-commerce"
