+++ src/main/java/kr/hhplus/be/ecommerce/EcommerceApplication.java
@@ -1,13 +1,13 @@
-package kr.hhplus.be.server;
+package kr.hhplus.be.ecommerce;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class ServerApplication {
+public class EcommerceApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(ServerApplication.class, args);
+		SpringApplication.run(EcommerceApplication.class, args);
 	}
 
 }
