+++ src/test/java/kr/hhplus/be/ecommerce/TestcontainersConfiguration.java
@@ -1,9 +1,6 @@
-package kr.hhplus.be.server;
+package kr.hhplus.be.ecommerce;
 
 import jakarta.annotation.PreDestroy;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.testcontainers.containers.MySQLContainer;
 import org.testcontainers.utility.DockerImageName;
